+++ src/lib/Components/Bidding/__stories__/BidFlow.story.tsx
@@ -7,12 +7,13 @@ import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import createEnvironment from "../../../relay/createEnvironment"
 
 import { Flex } from "../Elements/Flex"
-import { Serif16 } from "../Elements/Typography"
+import { Sans12, Serif14, Serif16 } from "../Elements/Typography"
 
 import { BiddingThemeProvider } from "../Components/BiddingThemeProvider"
 import { Checkbox } from "../Components/Checkbox"
 
 import BidFlow from "../../../Containers/BidFlow"
+import { Input } from "../Components/Input"
 import { BidResult } from "../Screens/BidResult"
 import { ConfirmBidScreen } from "../Screens/ConfirmBid"
 import { ConfirmFirstTimeBid } from "../Screens/ConfirmFirstTimeBid"
@@ -100,32 +101,57 @@ storiesOf("Bidding")
     return <BidResult winning={false} />
   })
 
-storiesOf("App Style/Input").add("Check Boxes", () => (
-  <BiddingThemeProvider>
-    <Flex mt={7}>
-      <Checkbox pl={3} pb={1}>
-        <Serif16 mt={2}>Remember me</Serif16>
-      </Checkbox>
+storiesOf("App Style/Input")
+  .add("Text Input", () => (
+    <BiddingThemeProvider>
+      <Flex mt={7} ml={4} mr={4}>
+        <Serif16 mb={2}>Title</Serif16>
+        <Input placeholder="Placeholder" mb={5} />
 
-      <Checkbox pl={3} pb={1} checked>
-        <Serif16 mt={2}>Remember me</Serif16>
-      </Checkbox>
+        <Serif16>Title</Serif16>
+        <Serif14 mb={2} color="black60">
+          Short description
+        </Serif14>
+        <Input placeholder="Placeholder" value="Content" mb={5} />
 
-      <Checkbox pl={3} pb={1} error="Please agree to the terms and conditions">
-        <Serif16 mt={2}>Agree to Terms and Conditions</Serif16>
-      </Checkbox>
+        <Input placeholder="Without Title" mb={5} />
 
-      <Checkbox p={3} checked error="Please agree to the terms and conditions">
-        <Serif16 mt={2}>Agree to Terms and Conditions</Serif16>
-      </Checkbox>
+        <Serif16 mb={2}>Error</Serif16>
+        <Input error placeholder="Placeholder" mb={3} />
+        <Sans12 color="red100">Error message</Sans12>
+      </Flex>
+    </BiddingThemeProvider>
+  ))
+  .add("Check Boxes", () => (
+    <BiddingThemeProvider>
+      <Flex mt={7}>
+        <Checkbox pl={3} pb={1}>
+          <Serif16 mt={2}>Remember me</Serif16>
+        </Checkbox>
 
-      <Checkbox pl={3} pb={1} disabled>
-        <Serif16 mt={2}>Remember me</Serif16>
-      </Checkbox>
+        <Checkbox pl={3} pb={1} checked>
+          <Serif16 mt={2}>Remember me</Serif16>
+        </Checkbox>
 
-      <Checkbox pl={3} pb={1} checked disabled>
-        <Serif16 mt={2}>Remember me</Serif16>
-      </Checkbox>
-    </Flex>
-  </BiddingThemeProvider>
-))
+        <Checkbox pl={3} pb={1} error>
+          <Serif16 mt={2} color="red100">
+            Agree to Terms and Conditions
+          </Serif16>
+        </Checkbox>
+
+        <Checkbox pl={3} pb={1} checked error>
+          <Serif16 mt={2} color="red100">
+            Agree to Terms and Conditions
+          </Serif16>
+        </Checkbox>
+
+        <Checkbox pl={3} pb={1} disabled>
+          <Serif16 mt={2}>Remember me</Serif16>
+        </Checkbox>
+
+        <Checkbox pl={3} pb={1} checked disabled>
+          <Serif16 mt={2}>Remember me</Serif16>
+        </Checkbox>
+      </Flex>
+    </BiddingThemeProvider>
+  ))