+++ CHANGELOG.md
@@ -44,6 +44,8 @@
 * Creates and verifies bidder position - sepans
 * Adds `<MarkdownRenderer>` component - sepans
 * Expands ConfirmBid tests - ash
+* Adds a `<Input>` component - yuki24
+* Adds a billing address screen - yuki24
 
 ### 1.4.6